the e is not int rather than unsigned long

+++ collatz4.c
@@ -43,7 +43,7 @@ void init_lut()
 void check(unsigned long n)
 {
 	do {
-		unsigned long e;
+		int e;
 
 		n >>= __builtin_ctzl(n);
